made a copy of squareScreen w/o using reducers
