+++ src/screens/SquareScreen.js
@@ -1,49 +1,29 @@
-import React, {useState} from 'react'
+import React, { useReducer } from 'react'
 import {View, Text, StyleSheet} from 'react-native'
 import ColorCounter from '../components/ColorCounter'
+
+// useReducer - Hook, Add some functionality to a component.
 
 const COLOR_INCREMENT = 15;
 
 const SquareScreen = () => {
-    const [red,
-        setRed] = useState(0)
-    const [green,
-        setGreen] = useState(0)
-    const [blue,
-        setBlue] = useState(0)
-
-    const setColor = (color, change) => {
-        // color === 'red', 'green', 'blue' change === +15 || -15
-        switch (color) {
-            case 'red':
-                red + change > 255 || red + change < 0 ? null : setRed(red + change);
-                return;
-            case 'green':
-                green + change > 255 || green + change < 0 ? null : setGreen(green + change);
-                return;
-            case 'blue':
-                blue + change > 255 || blue + change < 0 ? null : setBlue(blue + change);
-                return;
-            default:
-                return;
-        };
-    }
+    const [state, dispatch] = useReducer(reducer, {red: 0, green: 0, blue: 0})
 
     return (
         <View>
             <ColorCounter
-                onIncrease={() => setColor('red', COLOR_INCREMENT)}
-                onDecrease={() => setColor('red', -1 * COLOR_INCREMENT)}
+                onIncrease={() => }
+                onDecrease={() => }
                 color="red"/>
 
             <ColorCounter
-                onIncrease={() => setColor('green', COLOR_INCREMENT)}
-                onDecrease={() => setColor('green', -1 * COLOR_INCREMENT)}
+                onIncrease={() => }
+                onDecrease={() => }
                 color="green"/>
 
             <ColorCounter
-                onIncrease={() => setColor('blue', COLOR_INCREMENT)}
-                onDecrease={() => setColor('blue', -1 * COLOR_INCREMENT)}
+                onIncrease={() => }
+                onDecrease={() => }
                 color="blue"/>
 
             <View
